MultiDrawIndirect Mesh-Shading first steps.
Vulkan:
-Filling my renderer with some new extensions
-Filled VMA create info with some powerful flags(extensions from device).
-Now I can query allocation stats.
-CommandPools now are reset per-frame(best practices).
-Added VulkanDeviceManager, so it can choose best fit device. In future will be rewritten via base classes DeviceManager, Device, for multiple device rendering.
-Also every new descriptor set from shader has debug name now.
-Fixed some things from validation layers best practices, still left a bit of errors.
Renderer:
-Added CommandBufferSpecification, still WIP, im gonna move from fences, and let them stay only in swapchain, timeline semaphores are universal case. (can wait on host or device).
-Stripped out some logic of Mesh into MeshManager, so it can be extended in future with some kind of custom engine mesh compression. Importer WIP.
BindlessRenderer:
-Moved Camera UBO to Camera BDA.
Other:
-Integrating entt, ECS, Entities, Scenes, Components WIP, along RTX.
-Moved PackVec4ToU8Vec4() to Math.h
-Added s_MainThreadID in Threading.h
-Added SetIconImage in Window.h
-UILayer now can block events.

+++ Pathfinder/Source/Renderer/Renderer2D.h
@@ -49,7 +49,7 @@ class Renderer2D final : private Uncopyable, private Unmovable
 
         std::vector<Sprite> Sprites;
 
-        uint32_t FrameIndex = 0;
+        uint8_t FrameIndex = 0;
 
         static constexpr uint32_t s_MAX_QUADS              = 2500;
         static constexpr uint32_t s_MAX_VERTICES           = s_MAX_QUADS * 4;
